// TrippleWebPortal/src/app/core/models/locationModel.ts
export class locationModel {
    locationName: string;
    address: string;
    longitude: number;
    latitude: number
}

export class locationObject {
    id: string;
    userId: string;
    name: string;
    addressDisplay: string;
    addressPlace: string;
    autocomplete: boolean;
    x: number;
    y: number;
}